feat: reorganize repository structure and improve project organization
- Move cmd/logging-demo to examples/intermediate/07-structured-logging/
  - Add comprehensive README with usage examples and sample output
  - Create dedicated go.mod for proper module isolation
  - Integrate with progressive example learning path

- Clean up tmp/ directory and improve .gitignore
  - Remove build artifacts and temporary files from version control
  - Add patterns for tmp/, build/, and common Go artifacts
  - Prevent future temporary file commits

- Move test-phase3 to internal/integration-tests/
  - Separate internal testing infrastructure from user examples
  - Maintain proper project structure boundaries
  - Preserve integration test functionality

- Update documentation
  - Add structured logging example to README examples section
  - Maintain consistency with existing example numbering
  - Improve discoverability of production logging patterns

This reorganization creates cleaner separation between:
- User-facing examples (examples/)
- Internal testing code (internal/)
- Temporary build artifacts (excluded from git)

Resolves repository structure issues and improves maintainability.
