+++ examples/intermediate/07-structured-logging/go.mod
@@ -1,0 +1,7 @@
+module example07
+
+go 1.24.4
+
+require github.com/c0deZ3R0/go-sync-kit v0.1.0
+
+replace github.com/c0deZ3R0/go-sync-kit => ../../..
