--- test-phase3/go.mod
@@ -1,7 +1,0 @@
-module test-phase3
-
-go 1.24.4
-
-replace github.com/c0deZ3R0/go-sync-kit => ..
-
-require github.com/c0deZ3R0/go-sync-kit v0.0.0-00010101000000-000000000000
